+++ test_tf2/test/test_utils.cpp
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 #include <gtest/gtest.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 #include <tf2/utils.h>
 #include <tf2_kdl/tf2_kdl.h>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 #include <ros/time.h>
 
 double epsilon = 1e-9;
